+++ end-to-end-test.sh
@@ -5,8 +5,10 @@ set -euf -o pipefail
 enabled_collectors=$(cat << COLLECTORS
   arp
   bcache
+  bonding
   btrfs
   buddyinfo
+  cgroups
   conntrack
   cpu
   cpufreq
@@ -33,20 +35,22 @@ enabled_collectors=$(cat << COLLECTORS
   nfs
   nfsd
   pressure
+  processes
   qdisc
   rapl
   schedstat
+  selinux
+  slabinfo
   sockstat
   stat
-  thermal_zone
+  sysctl
   textfile
-  bonding
-  udp_queues 
+  thermal_zone
+  udp_queues
   vmstat
   wifi
   xfs
   zfs
-  processes
   zoneinfo
 COLLECTORS
 )
@@ -68,6 +72,20 @@ arch="$(uname -m)"
 case "${arch}" in
   aarch64|ppc64le) fixture='collector/fixtures/e2e-64k-page-output.txt' ;;
   *) fixture='collector/fixtures/e2e-output.txt' ;;
+esac
+
+# Only test CPU info collection on x86_64.
+case "${arch}" in
+  x86_64)
+    cpu_info_collector='--collector.cpu.info'
+    cpu_info_bugs='^(cpu_meltdown|spectre_.*|mds)$'
+    cpu_info_flags='^(aes|avx.?|constant_tsc)$'
+    ;;
+  *)
+    cpu_info_collector='--no-collector.cpu.info'
+    cpu_info_bugs=''
+    cpu_info_flags=''
+    ;;
 esac
 
 keep=0; update=0; verbose=0
@@ -104,17 +122,24 @@ fi
   --path.rootfs="collector/fixtures" \
   --path.procfs="collector/fixtures/proc" \
   --path.sysfs="collector/fixtures/sys" \
+  --path.udev.data="collector/fixtures/udev/data" \
   $(for c in ${enabled_collectors}; do echo --collector.${c}  ; done) \
   $(for c in ${disabled_collectors}; do echo --no-collector.${c}  ; done) \
   --collector.textfile.directory="collector/fixtures/textfile/two_metric_files/" \
   --collector.wifi.fixtures="collector/fixtures/wifi" \
   --collector.qdisc.fixtures="collector/fixtures/qdisc/" \
+  --collector.arp.device-exclude="nope" \
   --collector.netclass.ignored-devices="(dmz|int)" \
   --collector.netclass.ignore-invalid-speed \
   --collector.bcache.priorityStats \
-  --collector.cpu.info \
-  --collector.cpu.info.flags-include="^(aes|avx.?|constant_tsc)$" \
-  --collector.cpu.info.bugs-include="^(cpu_meltdown|spectre_.*|mds)$" \
+  "${cpu_info_collector}" \
+  --collector.cpu.info.bugs-include="${cpu_info_bugs}" \
+  --collector.cpu.info.flags-include="${cpu_info_flags}" \
+  --collector.stat.softirq \
+  --collector.sysctl.include="kernel.threads-max" \
+  --collector.sysctl.include="fs.file-nr" \
+  --collector.sysctl.include="fs.file-nr:total,current,max" \
+  --collector.sysctl.include-info="kernel.seccomp.actions_avail" \
   --web.listen-address "127.0.0.1:${port}" \
   --log.level="debug" > "${tmpdir}/node_exporter.log" 2>&1 &